Refactor elgamal and module
