+++ sinkhole-core/src/lib.rs
@@ -5,6 +5,5 @@ extern crate elgamal_ristretto;
 extern crate hex;
 extern crate rand_core;
 
-pub mod elgamal;
 pub mod errors;
 pub mod traits;
