+++ sinkhole-core/src/traits.rs
@@ -17,4 +17,18 @@ pub mod core {
         /// Returns the data for a given ID.
         fn retrieve(&self, query: Vec<Ciphertext>) -> Result<Ciphertext, StorageError>;
     }
+
+    pub trait Query {
+        /// Constructs an encrypted query.
+        fn new(
+            content: &[u8],
+            //public_key: sinkhole_core::PublicKey
+        ) -> &[u8];
+
+        /// Decrypts a decodes the result returned by the PIR server.
+        fn retrieve(
+            result: &[u8],
+            //private_key: sinkhole_core::PrivateKey
+        ) -> &[u8];
+    }
 }
