+++ sinkhole-elgamal/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 
-name = "sinkhole"
+name = "sinkhole-elgamal"
 version = "0.0.1"
 license = "MIT"
 authors = [ "Dean Eigenmann <dean@eigenmann.me>", "Gonçalo Pestana <g6pestana@gmail.com>" ]
@@ -14,7 +14,13 @@ Rust implementation of sinkhole PIR
 travis-ci = { repository = "hashmatter/sinkhole-rs" }
 
 [lib]
-name = "sinkhole"
+name = "sinkhole_elgamal"
 path = "src/lib.rs"
 
-## @TODO THIS PACKAGE WOULD PROBABLY DO EVERYTHING?
+[dependencies]
+sinkhole-core = { path = "../sinkhole-core" }
+elgamal_ristretto = "0.2.3"
+curve25519-dalek = "2.0.0"
+rand_core = "0.5.1"
+bincode = "1.2.1"
+
