+++ sinkhole-elgamal/README.md
@@ -1,0 +1,5 @@
+# ElGamal implementation of sinkhole-rs 🦀
+
+Implementation of the sinhole-rs PIR protocol using ElGamal over the Ristretto
+curve cryptosystem.
+
