+++ sinkhole-elgamal/src/lib.rs
@@ -1,0 +1,7 @@
+extern crate curve25519_dalek;
+extern crate elgamal_ristretto;
+extern crate rand_core;
+extern crate sinkhole_core;
+
+pub mod client;
+pub mod storage;
