--- sinkhole-query/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-
-name = "sinkhole-query"
-version = "0.0.1"
-license = "MIT"
-authors = [ "Dean Eigenmann <dean@eigenmann.me>", "Gonçalo Pestana <g6pestana@gmail.com>" ]
-repository = "https://github.com/hashmatter/sinkhole-rs"
-description = """
-The query generation and decryption used by the `sinkhole` library.
-"""
-
-[badges]
-travis-ci = { repository = "hashmatter/sinkhole-rs" }
-
-[dependencies]
-curve25519-dalek = { version = "2.0.0", default-features = false, features = ["serde"] }
-elgamal_ristretto = "0.2.3"
-
